Moved Template function definition in header
* pro: no implicit declaration
* contra: more compile time
