+++ source/book/include/marlin/book/BookStore.h
@@ -5,6 +5,9 @@
 #include <string>
 #include <string_view>
 #include <unordered_map>
+#include <typeindex>
+#include <typeinfo>
+#include <stdexcept>
 
 // -- Marlin includes
 #include "marlin/Exceptions.h"
@@ -12,6 +15,12 @@
 // -- MarlinBook includes
 #include "marlin/book/ROOTAdapter.h"
 #include "marlin/book/Entry.h"
+#include "marlin/book/Condition.h"
+#include "marlin/book/EntryData.h"
+#include "marlin/book/Flags.h"
+#include "marlin/book/Hist.h"
+#include "marlin/book/MemLayout.h"
+#include "marlin/book/Selection.h"
 
 namespace marlin {
   /// contains classes needed to book and managed booked objects.
@@ -75,7 +84,7 @@ namespace marlin {
       Entry &get( const EntryKey &key ) { 
         try {
           return _entries[key.hash]; 
-        } catch(const auto&) {
+        } catch(const std::out_of_range&) {
           MARLIN_THROW_T(BookStoreException, "Invalid key.");
         }
       
@@ -157,6 +166,99 @@ namespace marlin {
       /// stores Entries created by BookStore.
       std::vector< Entry > _entries{} ;
     } ;
+
+
+    //--------------------------------------------------------------------------
+
+
+    void BookStore::addEntry( const std::shared_ptr< EntryBase > &entry,
+                              EntryKey &                          key ) {
+      EntryKey k = key ;
+      k.hash     = _entries.size() ;
+      _entries.push_back( Entry( entry, k ) ) ;
+    }
+
+    //--------------------------------------------------------------------------
+
+    template < class T, typename... Args_t >
+    EntrySingle< T > BookStore::bookSingle( const std::string_view &path,
+                                            const std::string_view &name,
+                                            Args_t... ctor_p ) {
+      EntryKey key{std::type_index( typeid( T ) )} ;
+      key.name  = name ;
+      key.path  = path ;
+      key.mInstances   = 1 ;
+      key.flags = Flags::Book::Single ;
+
+      auto entry = std::make_shared< EntrySingle< T > >( Context(
+        std::make_shared< SingleMemLayout< T, Args_t... > >( ctor_p... ) ) ) ;
+
+      addEntry( entry, key ) ;
+
+      return *std::static_pointer_cast< const EntrySingle< T > >( entry ) ;
+    }
+
+    //--------------------------------------------------------------------------
+
+    template < class T, typename... Args_t >
+    EntryMultiCopy< T > BookStore::bookMultiCopy( std::size_t             n,
+                                                  const std::string_view &path,
+                                                  const std::string_view &name,
+                                                  Args_t... ctor_p ) {
+      EntryKey key{std::type_index( typeid( T ) )} ;
+      key.name   = name ;
+      key.path   = path ;
+      key.mInstances    = n ;
+      key.flags  = Flags::Book::MultiCopy ;
+      auto entry = std::make_shared< EntryMultiCopy< T > >( Context(
+        std::make_shared< SharedMemLayout< T, trait< T >::Merge, Args_t... > >(
+          n, ctor_p... ) ) ) ;
+
+      addEntry( entry, key ) ;
+
+      return *std::static_pointer_cast< const EntryMultiCopy< T > >( entry ) ;
+    }
+
+    Selection BookStore::find( const Condition &cond ) {
+      return Selection::find( _entries.cbegin(), _entries.cend(), cond ) ;
+    }
+
+    //--------------------------------------------------------------------------
+
+    void BookStore::remove( const Entry &e ) { get( e.key() ).clear(); }
+
+    //--------------------------------------------------------------------------
+
+    void BookStore::remove( const Selection &selection ) {
+      for ( const Entry &e : selection ) {
+        remove( e ) ;
+      }
+    }
+
+    //--------------------------------------------------------------------------
+
+    void BookStore::clear() { _entries.resize( 0 ); }
+
+    //--------------------------------------------------------------------------
+
+    template < class T, typename... Args_t >
+    EntryMultiShared< T >
+    BookStore::bookMultiShared( const std::string_view &path,
+                                const std::string_view &name,
+                                Args_t... ctor_p ) {
+      EntryKey key{std::type_index( typeid( T ) )} ;
+      key.name  = name ;
+      key.path  = path ;
+      key.mInstances = 1 ;
+      key.flags = Flags::Book::MultiShared ;
+
+      auto entry = std::make_shared< EntryMultiShared< T > >( Context(
+        std::make_shared< SingleMemLayout< T, Args_t... > >( ctor_p... ) ) ) ;
+
+      addEntry( entry, key ) ;
+
+      return *std::static_pointer_cast< const EntryMultiShared< T > >( entry ) ;
+    }
 
   } // end namespace book
